imprime tanque na tela

+++ Coordenada.cpp
@@ -4,7 +4,7 @@ Coordenada::Coordenada()
 {
 }
 
-Coordenada::Coordenada(const unsigned short int x, const unsigned short int y):x{x}, y{y}      
+Coordenada::Coordenada(const unsigned int x, const unsigned int y):x{x}, y{y}      
 {
 }
 
@@ -12,18 +12,18 @@ Coordenada::~Coordenada()
 {
 }
 
-const unsigned short int Coordenada::getX() const {
+const unsigned int Coordenada::getX() const {
     return this->x;
 }
-void Coordenada::setX(const unsigned short int x) {
+void Coordenada::setX(const unsigned int x) {
 
     this->x = x;
 }
 
-const unsigned short int Coordenada::getY() const {
+const unsigned int Coordenada::getY() const {
     return this->y;
 }
 
-void Coordenada::setY(const unsigned short int y){
+void Coordenada::setY(const unsigned int y){
     this->y = y;
 }
